# docs/sa/API_Specification.yaml
openapi: 3.0.3
info:
  title: ServiceDesk CRM API
  description: |
    REST API для системы автоматизации сервисной компании по обслуживанию банкоматов и платёжного оборудования.

    ## Аутентификация
    Все защищённые эндпоинты требуют заголовок `Authorization: Bearer <token>`.
    Токен получается через `POST /auth/login`. Срок действия — 480 минут (8 часов).

    ## Формат ошибок
    ```json
    { "error": "ERROR_CODE", "message": "Описание ошибки", "details": {} }
    ```

    ## Пагинация
    Поддерживается параметрами `?page=1&size=20`.
    Ответ всегда включает `total`, `page`, `size`, `pages`.
  version: "2.0.0"
  contact:
    name: Solution Architect
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: http://localhost/api/v1
    description: Docker Compose (via Nginx)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: TICKET_NOT_FOUND
        message:
          type: string
          example: "Заявка #42 не найдена"
        details:
          type: object
      required: [error, message, details]
    PaginatedMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
          description: Текущая страница (начиная с 1)
        size:
          type: integer
          description: Размер страницы
        pages:
          type: integer
          description: Общее количество страниц
      required: [total, page, size, pages]

    # ─── Auth ─────────────────────────────────────────────────────────────────
...
    SparePartResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
          nullable: true
        stock_quantity:
          type: integer
        unit:
          type: string
        min_stock:
          type: integer
        price:
          type: number
          format: float
          nullable: true
      required: [id, name, stock_quantity, unit, min_stock]

    # ─── File ─────────────────────────────────────────────────────────────────
...
    NotificationResponse:
      type: object
      properties:
        id:
          type: integer
        event_type:
          type: string
        message:
          type: string
        entity_type:
          type: string
          nullable: true
        entity_id:
          type: integer
          nullable: true
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, event_type, message, is_read, created_at]
...
    NotificationSettingResponse:
      type: object
      properties:
        event_type:
          type: string
        channel:
          type: string
          enum: [email, telegram, in_app]
        enabled:
          type: boolean
      required: [event_type, channel, enabled]
    NotificationSettingUpdate:
      type: object
      properties:
        event_type:
          type: string
        channel:
          type: string
          enum: [email, telegram]
          description: "in_app нельзя отключить (BR-F-1400)"
        enabled:
          type: boolean
      required: [event_type, channel, enabled]
  responses:
...
    NotFound:
      description: Запись не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BusinessError:
      description: Нарушение бизнес-правила
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - BearerAuth: []
paths:
...
  /parts:
    get:
      tags: [Запчасти]
      summary: Список запчастей
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: low_stock
          schema:
            type: boolean
          description: Только позиции с остатком ниже минимума
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/SparePartResponse'

    post:
      tags: [Запчасти]
      summary: Добавить запчасть в номенклатуру
      description: admin, warehouse.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SparePartResponse'
      responses:
        '201':
          description: Создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePartResponse'
  /parts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Запчасти]
      summary: Получить запчасть
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePartResponse'

    put:
      tags: [Запчасти]
      summary: Обновить запчасть
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SparePartResponse'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePartResponse'
...
  /notifications:
    get:
      tags: [Уведомления]
      summary: Список уведомлений текущего пользователя
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/NotificationResponse'
...
  /notifications/{id}/read:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    post:
      tags: [Уведомления]
      summary: Отметить уведомление прочитанным
      responses:
        '204':
          description: OK
        '404':
          $ref: '#/components/responses/NotFound'
...
  /notifications/settings:
    get:
      tags: [Уведомления]
      summary: Настройки уведомлений текущего пользователя
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationSettingResponse'

    put:
      tags: [Уведомления]
      summary: Обновить настройки уведомлений
      description: in_app нельзя отключить (BR-F-1400).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/NotificationSettingUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationSettingResponse'
        '400':
          $ref: '#/components/responses/BusinessError'
...